Updated index.html
I tried making the format nicer in the About section because I think it looks to linear, but it was looking weird when you changed the size of the web page, so I'll have to look into that more. I posted that one on a different branch. This file just updated a social media section and i rearranged some pictures.